+++ common/types.go
@@ -1,7 +1,7 @@
 package common
 
 type Finfo struct {
-	NameLength int `struc:"int16,sizeof=Name"`
+	NameLength int64 `struc:"int16,sizeof=Name"`
 	Name       string
 	Size       int64
 	IsDir      bool
